requirements.txt
# Core Trading Dependencies
MetaTrader5>=5.0.37
pandas>=1.5.0
numpy>=1.21.0
TA-Lib>=0.4.25

# Web Framework
streamlit>=1.28.0

# HTTP Requests for Telegram
requests>=2.28.0

# Date and Time handling
python-dateutil>=2.8.2

# Environment Variables
python-dotenv>=0.19.0

# Data Visualization (for ICT charts)
plotly>=5.14.0

# Additional Scientific Computing
scipy>=1.9.0

# Timezone handling
pytz>=2022.1

# Network timeout handling
urllib3>=1.26.12

# Retry mechanisms
tenacity>=8.1.0

# Progress bars
tqdm>=4.64.0

# Color output in terminal
colorama>=0.4.5